# backend/storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tinder Clone API",
    "version": "1.0.0",
    "description": "API for a Tinder-like dating application"
  },
  "servers": [
    {
      "url": "http://192.168.2.53:8000/api/v1",
      "description": "Development Server"
    }
  ],
  "paths": {
...
    "/likes/like": {
      "post": {
        "tags": ["Likes"],
        "summary": "Like a person",
        "operationId": "like",
        "security": [
          {
            "sanctum": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["person_id"],
                "properties": {
                  "person_id": {
                    "type": "integer",
                    "description": "ID of the person to like"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Like recorded successfully"
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/likes/dislike": {
      "post": {
        "tags": ["Likes"],
        "summary": "Dislike a person",
        "operationId": "dislike",
        "security": [
          {
            "sanctum": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["person_id"],
                "properties": {
                  "person_id": {
                    "type": "integer",
                    "description": "ID of the person to dislike"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dislike recorded successfully"
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/likes/liked-people": {
      "get": {
        "tags": ["Likes"],
        "summary": "Get liked people",
        "operationId": "likedPeople",
        "security": [
          {
            "sanctum": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of liked people",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      },
                      "age": {
                        "type": "integer"
                      },
                      "location": {
                        "type": "string"
                      },
                      "bio": {
                        "type": "string"
                      },
                      "pictures": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sanctum": {
        "type": "apiKey",
        "description": "Enter token in format (Bearer <token>)",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}